backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
openai==1.3.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
twilio==8.10.0
boto3==1.29.0
requests==2.31.0
pandas==2.1.2
python-csv==0.0.13
